+++ assets/javascripts/kitten/index.js
@@ -198,6 +198,7 @@ export { SpeechBubbleIcon } from './components/graphics/icons/speech-bubble-icon
 
 // Graphics/IconsNext
 export { AirplaneIconNext } from './components/graphics/icons-next/airplane-icon-next'
+export { BankCardIconNext } from './components/graphics/icons-next/bank-card-icon-next'
 export { BubbleIconNext } from './components/graphics/icons-next/bubble-icon-next'
 export { CheckedShieldIconNext } from './components/graphics/icons-next/checked-shield-icon-next'
 export { ChronoIconNext } from './components/graphics/icons-next/chrono-icon-next'
@@ -205,6 +206,7 @@ export { CogIconNext } from './components/graphics/icons-next/cog-icon-next'
 export { ColorCheckedShieldIconNext } from './components/graphics/icons-next/color-checked-shield-icon-next'
 export { CrossCircleIconNext } from './components/graphics/icons-next/cross-circle-icon-next'
 export { CrossIconNext } from './components/graphics/icons-next/cross-icon-next'
+export { CubeIconNext } from './components/graphics/icons-next/cube-icon-next'
 export { EditIconNext } from './components/graphics/icons-next/edit-icon-next'
 export { EnvelopeIconNext } from './components/graphics/icons-next/envelope-icon-next'
 export { EyeIconNext } from './components/graphics/icons-next/eye-icon-next'